requirements.txt
fastapi
uvicorn
tortoise-orm
httpx
python-dotenv
python-weather
requests
asyncpg
pydantic
pydantic_settings